Refactor: Move functions to utils module
- Separated chat, log, UI, utility, and vector store functions from app.py.
- Created new files in the utils directory:
  - utils/chat.py
  - utils/logs.py
  - utils/ui.py
  - utils/utils.py
  - utils/vectorstore.py
- Updated app.py to import these functions from the new utils module.
