+++ app.py
@@ -1,294 +1,14 @@
-import os, tempfile
 import streamlit as st
-import json
 from streamlit import session_state as sst
-from langchain_community.document_loaders.pdf import PyPDFLoader
-from langchain.indexes import VectorstoreIndexCreator
-from langchain_openai import ChatOpenAI, OpenAIEmbeddings
-from langchain_community.vectorstores import FAISS
-from streamlit_chat import message
-from datetime import datetime
-from io import BytesIO
-from pathlib import Path
+from langchain_openai import ChatOpenAI
+from utils.chat import initialize_chat_history, show_chat, add_to_chat
+from utils.logs import initialize_log, display_log, add_to_log
+from utils.vectorstore import get_vectorstore
+from utils.ui import navbar
+from utils.utils import load_css, prepare_download_file
 
 OPENAI_API_KEY = st.secrets['OPENAI_API_KEY']
 
-def get_timestamp():
-  """Generates and returns a timestamp for the current time
-
-  Returns:
-      datetime: Current time in format of Hour:Minute:Second
-  """
-  return datetime.now().strftime("%H:%M:%S")
-
-def initialize_log():
-  """
-  Initializes log message list
-  """
-  sst["log"] = [{
-    "time" : get_timestamp(),
-    "status" : "info",
-    "message" : "Displaying background activity.."
-  }]
-
-def display_log(logs:list):
-  """
-  Display a log message in the sidebar when `show_bts` is active.
-
-  Args:
-      logs (list): List of log messages
-  """
-  for log_msg in logs:
-    if log_msg.get('status') == "info":
-      sst.container.caption(f":orange[[{log_msg.get('time')}]] {log_msg.get('message')}")
-    elif log_msg.get('status') == "success":
-      sst.container.caption(f":orange[[{log_msg.get('time')}]] :green[{log_msg.get('message')}]")
-    elif log_msg.get('status') == "error":
-      sst.container.caption(f":orange[[{log_msg.get('time')}]] :red[{log_msg.get('message')}]")
-
-def add_to_log(message:str, status="info"):
-  """Adds message to List of log messages
-
-  Args:
-      message (str): log message
-      status (str, optional): "info", "success" or "error". Status of log message. Defaults to "info".
-  """
-  # Check if show_bts exists in session state, default to False if not
-  show_bts = getattr(sst, 'show_bts', False)
-  
-  if show_bts:
-    log_entry = {
-      "time" : get_timestamp(),
-      "status" : status,
-      "message" : message
-    }
-    sst.log.insert(0, log_entry)
-    sst.container.caption(f":orange[[now]] :grey-background[{message}]")
-  print(f"[{get_timestamp()}] : {message}")
-    
-def initialize_chat_history():
-  """
-  Initialize chat history with a welcome message from AskNotes.ai.
-  """
-  add_to_log("Initializing Chat History.")
-  sst["chat_history"] = [
-    {
-      'role': 'ai',
-      'content': "Hi! I'm AskNotes.ai. Ask me anything about the uploaded PDF!"
-    }
-  ]
-  add_to_log("Chat History Initialized.", "success")
-
-def show_chat(messages: list):
-  """
-  Display chat messages stored in session state.
-
-  Args:
-      messages (list): List of messages in the chat history.
-  """
-  for i, msg in enumerate(messages):
-    message(
-      message=msg['content'], 
-      is_user=msg['role'] == 'user', 
-      key=str(i)
-    )
-  add_to_log("Displaying Chat.")
-
-def add_to_chat(role, content):
-  """
-  Add a message to the chat history and display it.
-
-  Args:
-      role (str): 'user' or 'ai' to indicate message origin.
-      content (str): Text content of the message.
-  """
-  sst.chat_history.append(
-    {
-      "role": role, 
-      "content": content
-    }
-  )
-  add_to_log("Message Added to Chat History..")
-  
-  message(
-    message=content, 
-    is_user=(role == 'user')
-  )
-  add_to_log("Displaying Message..")
-
-def prepare_download_file(format_type):
-  
-    """
-    Prepare the chat history file based on the selected format.
-
-    Args:
-        format_type (str): 'JSON' or 'TXT'.
-
-    Returns:
-        tuple: Tuple containing the file content, file name, and file type.
-        
-    """
-    if format_type == "JSON":
-      
-        add_to_log("Preparing chat history as JSON...")
-        return BytesIO(json.dumps(sst.chat_history, indent=4).encode('utf-8')), "chat_history.json", "application/json"
-    
-    elif format_type == "TXT":
-      
-        add_to_log("Preparing chat history as TXT...")
-        text_output = "\n\n--------------------------------------------------------------\n\n".join(
-            [
-                f"{'User' if msg['role'] == 'user' else 'AI'}: {msg['content']}"
-                for msg in sst.chat_history
-            ]
-        )
-        text_output = f"Chat History:\n{'=' * 40}\n\n{text_output}\n\n{'=' * 40}\nEnd of Chat"
-        return BytesIO(text_output.encode('utf-8')), "chat_history.txt", "text/plain"
-    
-    return None, None, None
-  
-def get_vectorstore():
-    """
-    Creates or retrieves an existing vectorstore from session state.
-    Includes enhanced error handling for PDFs with graphics or no text content.
-
-    Returns:
-        Vectorstore object stored in session state.
-    """
-    add_to_log("Creating Vectorstore..")
-
-    loader_list, temp_paths = get_loader(sst.pdf_files)
-    try:
-        with st.spinner("Creating Vectorstore..."):
-            if not loader_list:
-                # Clear PDF files from session state to allow fresh start
-                if "vectorstore" in sst:
-                    del sst.vectorstore
-                return None
-            
-            embeddings = OpenAIEmbeddings()
-            try:
-                sst.vectorstore = VectorstoreIndexCreator(
-                    vectorstore_cls=FAISS, 
-                    embedding=embeddings
-                ).from_loaders(loader_list)
-                add_to_log("Created Vectorstore Successfully..", "success")
-                return sst.vectorstore
-            except Exception as e:
-                st.toast(f"Error creating vectorstore. Try another PDF.", icon="⚠️")
-                add_to_log(f"Error: Unable to create vectorstore - {str(e)}", "error")
-                if "vectorstore" in sst:
-                    del sst.vectorstore
-                return None
-    except Exception as e:
-        st.toast("Unexpected error. Please try again with a different PDF.", icon="⚠️")
-        add_to_log(f"Error: {str(e)}", "error")
-        if "vectorstore" in sst:
-            del sst.vectorstore
-        return None
-    finally:
-        delete_temp_files(temp_paths)
-
-def get_loader(pdf_files: list):
-    """
-    Creates PDF loaders from uploaded PDFs, saving each file temporarily.
-    Includes enhanced error handling for PDFs with graphics.
-
-    Args:
-        pdf_files (list): List of uploaded PDFs.
-
-    Returns:
-        list: PDF loaders for processing.
-        list: Temporary file paths for cleanup.
-    """
-    add_to_log("Processing PDFs..")
-    pdf_loader_list = []
-    temp_paths = []
-    has_unreadable_pdfs = False
-    
-    try:
-        with st.spinner("Loading PDFs..."):
-            for pdf in pdf_files:
-                try:
-                    with tempfile.NamedTemporaryFile(delete=False, suffix=".pdf") as f:
-                        f.write(pdf.getvalue())
-                        temp_paths.append(f.name)
-                        # Create loader with enhanced error handling
-                        loader = PyPDFLoader(
-                            f.name,
-                            extract_images=False  # Skip image extraction to avoid errors
-                        )
-                        # Test load to catch any immediate issues
-                        try:
-                            pages = loader.load()
-                            # Only add loader if it successfully extracted text
-                            if any(len(page.page_content.strip()) > 0 for page in pages):
-                                pdf_loader_list.append(loader)
-                                add_to_log(f"Successfully processed {pdf.name}", "success")
-                            else:
-                                has_unreadable_pdfs = True
-                                add_to_log(f"No text content found in {pdf.name}", "error")
-                        except Exception as e:
-                            has_unreadable_pdfs = True
-                            add_to_log(f"Error loading pages from {pdf.name}", "error")
-                except Exception as e:
-                    add_to_log(f"Error processing {pdf.name}", "error")
-                    continue
-            
-            if pdf_loader_list:
-                add_to_log("PDFs loaded successfully!", "success")
-                if has_unreadable_pdfs:
-                    st.toast("Some PDFs were skipped. Only PDFs with readable text were included.", icon="ℹ️")
-                return pdf_loader_list, temp_paths
-            else:
-                add_to_log("No valid PDFs could be processed", "error")
-                st.toast("Please upload PDFs with readable text content.", icon="⚠️")
-                return None, temp_paths
-            
-    except Exception as e:
-        add_to_log(f"Error: {str(e)}", "error")
-        st.toast("Error loading PDFs. Please try again with different files.", icon="⚠️")
-        return None, temp_paths
-
-def delete_temp_files(temp_paths: list):
-  """
-  Deletes temporary files created during PDF loading.
-
-  Args:
-      temp_paths (list): Paths of temporary files to delete.
-  """
-  add_to_log("Deleting Temporary files..")
-  for path in temp_paths:
-    try:
-      os.remove(path)
-    except Exception as e:
-      st.warning(f"Failed to delete temp file {path}: {e}")
-      add_to_log("Error: Failed to delete temporary files..", "error")
-  add_to_log("Temporary Files Deleted.", "success")
-
-
-def load_css(path:Path= Path('static/styles.css')) -> str:
-  """
-  Loads a CSS stylesheet from a local file.
-
-  Args:
-      path (Path): Path to the CSS file (default: 'static/styles.css').
-
-  Returns:
-      str: The content of the CSS file if successfully loaded, otherwise an empty string.
-  """
-  # Load CSS stylesheet
-  try:
-    if not path.is_file():
-      add_to_log(f"❗Error loading stylesheet: {path} does not exist.", "error")
-      return ""
-    with open(path) as f:
-        custom_css = f.read()
-    return custom_css
-  except FileNotFoundError:
-    add_to_log("❗Error loading stylesheet: File not found.", "error")
-  except Exception as e:
-    add_to_log("❗Error loading stylesheet.", "error")
 
 def main():
     # Set up the main page layout and title
@@ -301,23 +21,7 @@ def main():
 
     # ---- Navbar ----
     with st.container():
-        app_col, home_col, about_col = st.columns(3)
-
-        app_col.page_link(
-            page='app.py',
-            label='App',
-            icon=':material/robot_2:'
-        )
-        home_col.page_link(
-            page= "pages/home.py",
-            label= "Home",
-            icon= ':material/home:'
-        )
-        about_col.page_link(
-            page='pages/about.py',
-            label='About',
-            icon=':material/star:',
-        )
+        navbar()
     st.markdown("---")
 
     # Title
